Proje2/requirements.txt
numpy
pandas
opencv-python
Pillow
scikit-learn
matplotlib
seaborn
streamlit
